// src/app/lib/generated/models/GeneratedImages.ts
export interface GeneratedImages {
    /**
     * 
     * @type {string}
     * @memberof GeneratedImages
     */
    type: GeneratedImagesTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof GeneratedImages
     */
    chatbotId: string;
    /**
     * 
     * @type {string}
     * @memberof GeneratedImages
     */
    chatbotName: string;
    /**
     * 
     * @type {string}
     * @memberof GeneratedImages
     */
    chatbotAvatar: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof GeneratedImages
     */
    urls: Array<string>;
    /**
     * 
     * @type {Date}
     * @memberof GeneratedImages
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof GeneratedImages
     */
    eta: string;
    /**
     * 
     * @type {ImageGenerationStatus}
     * @memberof GeneratedImages
     */
    status: ImageGenerationStatus;
    /**
     * 
     * @type {string}
     * @memberof GeneratedImages
     */
    generationId: string;
    /**
     * 
     * @type {boolean}
     * @memberof GeneratedImages
     */
    seenByUser: boolean;
}


/**
 * @export
 */
export const GeneratedImagesTypeEnum = {
    Liked: 'liked',
    Generated: 'generated'
} as const;
export type GeneratedImagesTypeEnum = typeof GeneratedImagesTypeEnum[keyof typeof GeneratedImagesTypeEnum];